Successfully reading in and using tilesheet, movement seems inversed, need to render player, but great success!

+++ src/Image_Utils.py
@@ -1,8 +1,11 @@
+import pygame
+from pygame.locals import *
+
 def load_tileset(file, image_width, image_height, tile_width, tile_height):
     image = pygame.image.load(file).convert_alpha()
     tileset = []
-    num_tiles_x = image_width/tile_width
-    num_tiles_y = image_height/tile_height
+    num_tiles_x = int(image_width/tile_width)
+    num_tiles_y = int(image_height/tile_height)
     
     for i in range(num_tiles_x):
         for j in range(num_tiles_y):
